update on bytes32

+++ test/abstract/OpTest.sol
@@ -90,8 +90,8 @@ abstract contract OpTest is RainterpreterExpressionDeployerNPE2DeploymentTest {
         IntegrityCheckStateNP memory integrityState = LibIntegrityCheckNP.newState("", 0, constants);
         (uint256 calcInputs, uint256 calcOutputs) = integrityFn(integrityState, operand);
         assertEq(calcInputs, inputs.length, "inputs length");
-        assertEq(calcInputs, (OperandV2.unwrap(operand) >> 0x10) & 0x0F, "operand inputs");
-        assertEq(calcOutputs, OperandV2.unwrap(operand) >> 0x14, "operand outputs");
+        assertEq(calcInputs, uint256((OperandV2.unwrap(operand) >> 0x10) & 0x0F), "operand inputs");
+        assertEq(calcOutputs, uint256(OperandV2.unwrap(operand) >> 0x14), "operand outputs");
         return calcOutputs;
     }
